% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difPlot-b.R
\name{difPlot}
\alias{difPlot}
\title{Differential Item Functioning IRT models.}
\usage{
difPlot(dat, theta = seq(-3, 3, 0.01), title = NULL, ddir = getwd(),
  save = FALSE, model, grp, fln = paste0(model, "_", "dif_", colnames(dat),
  ".jpg"), dpi = 300, height = 8.5, width = 10, itmNam = colnames(dat),
  silent = FALSE)
}
\arguments{
\item{dat}{A data frame containing at least one item as a column.}

\item{theta}{The ability parameter across which to plot response parameterization.}

\item{save}{Whether or not the generated plots will be saved. Default is \code{"FALSE"}}

\item{model}{The desired model to fit to the data.}

\item{type}{The plot type to generate. See Details for a list of valid types.}
}
\value{
Prints the specified plot by default leaving the options to save
	 to the user.
}
\description{
\code{difPlot} returns a plot within a new device and saves to chosen directory.
}
\details{
This function specializes in generating models to examine DIF.
Currently, the only supported type is \code{"icc"} with \code{"1PL"} and \code{"2PL"} models.
}
\examples{
## Example notation:
theta <- seq(-3,3, 0.01)
difPlot(data, theta, grp = group, model = "1PL", type = "icc")


}

