% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtPlot-b.R
\name{irtPlot}
\alias{irtPlot}
\title{Plot binary Item Response Theory models.}
\usage{
irtPlot(dat, theta = seq(-3, 3, 0.001), title = NULL, ddir = getwd(),
  save = FALSE, model, type, filename = paste0(model, "_", type, "_",
  ".jpg"), dpi = 800, height = 6.5, width = 10, itmNam = colnames(dat),
  subS = NULL)
}
\arguments{
\item{dat}{A data frame containing at least one item as a column.}

\item{theta}{The ability parameter across which to plot response parameterization.}

\item{save}{Whether or not the generated plots will be saved. Default is \code{"FALSE"}}

\item{model}{The desired model to fit to the data.}

\item{type}{The plot type to generate. See Details for a list of valid types.}
}
\value{
Prints the specified plot by default leaving the options to save
	 to the user.
}
\description{
\code{irtPlot} returns a plot within a new device and saves to chosen directory.
}
\details{
This function is capable of generating several models and plot types.
Namely, the models are \code{"1PL"},\code{"2PL"}, and \code{"3PL"} and the possible types are
\code{"icc"}, \code{"tcc"}, \code{"iif"}, \code{"tif"}, \code{"likl"}, and \code{"logl"}
}
\examples{
## Load data:
library(ltm)
data <- Science[,c(1,3)]
colnames(data) <- paste0("Item ", 1:ncol(data))

}

