% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyPlot-b.R
\name{polyPlot}
\alias{polyPlot}
\title{Plot Polytomous Item Response Theory models.}
\usage{
polyPlot(dat, theta = seq(-3, 3, 0.01), title = NULL, ddir = getwd(),
  save = FALSE, model, type, fln = paste0(model, "_", type, "_",
  colnames(dat), ".jpg"), dpi = 300, height = 8.5, width = 10,
  itmNam = colnames(dat), silent = FALSE)
}
\arguments{
\item{dat}{A data frame containing at least one item as a column.}

\item{theta}{The ability parameter across which to plot response parameterization.}

\item{save}{Whether or not the generated plots will be saved. Default is \code{"FALSE"}}

\item{model}{The desired model to fit to the data.}

\item{type}{The plot type to generate. See Details for a list of valid types.}
}
\value{
Prints the specified plot by default leaving the options to save
	 to the user.
}
\description{
\code{polyPlot} returns a plot within a new device and saves to chosen directory.
}
\details{
This function specializes in generating polytomous models and plot types.
Namely, the models are \code{"grm"} and \code{"pcm"} and the possible types are
\code{"icc"}, \code{"icf"}, and \code{"crp"}.
}
\examples{
## Example notation:
library(ltm)
data <- Science[,c(1,3)]
colnames(data) <- paste0("Item ", 1:ncol(data))
polyPlot(data, theta, model = "grm", type = "crp")
}

